pyproject.toml
[project]
name = "agentic-retail-demo"
version = "0.1.0"
description = "Agentic Retail Application Demo"
requires-python = ">=3.11.4"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "google-adk==1.18.0",
    "google-cloud-bigquery>=3.24.0",
    "google-cloud-storage>=2.17.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "aiohttp==3.13.2",
]

[tool.setuptools]
packages = ["app"]
